Extract `switchNewExpressionToCallExpression`
